refactor(cooklang): extract ReportPresenter, use shared service in ReportContribution
