+++ packages/cooklang/src/browser/cooklang-frontend-module.ts
@@ -37,6 +37,9 @@ import { MENU_PREVIEW_WIDGET_ID, createMenuPreviewWidget } from './menu-preview-
 import { MenuPreviewContribution } from './menu-preview-contribution';
 import { REPORT_WIDGET_ID, ReportWidgetOptions, createReportWidget } from './report-widget';
 import { ReportContribution } from './report-contribution';
+import { ReportConfigService } from './report-config-service';
+import { ReportPresenter } from './report-presenter';
+import { ReportWidgetPresenter } from './report-widget-presenter';
 import { bindCooklangPreferences } from '../common';
 
 export default new ContainerModule(bind => {
@@ -89,6 +92,11 @@ export default new ContainerModule(bind => {
         createWidget: (options: ReportWidgetOptions) =>
             createReportWidget(ctx.container, options),
     })).inSingletonScope();
+
+    // Report config + presenter (shared by the command and the AI render tool)
+    bind(ReportConfigService).toSelf().inSingletonScope();
+    bind(ReportWidgetPresenter).toSelf().inSingletonScope();
+    bind(ReportPresenter).toService(ReportWidgetPresenter);
 
     // Report command and context menu
     bind(ReportContribution).toSelf().inSingletonScope();
