+++ packages/cooklang/src/browser/report-contribution.ts
@@ -14,17 +14,16 @@
 import { injectable, inject } from '@theia/core/shared/inversify';
 import { CommandContribution, CommandRegistry, Command } from '@theia/core/lib/common/command';
 import { MenuModelRegistry, MenuContribution } from '@theia/core/lib/common/menu';
-import { ApplicationShell, Widget, WidgetManager } from '@theia/core/lib/browser';
 import { QuickPickService, QuickPickItem, QuickPickSeparator } from '@theia/core/lib/common/quick-pick-service';
 import { nls } from '@theia/core/lib/common/nls';
-import { NavigatableWidget } from '@theia/core/lib/browser/navigatable-types';
-import { EditorManager, EDITOR_CONTEXT_MENU } from '@theia/editor/lib/browser';
+import { EDITOR_CONTEXT_MENU } from '@theia/editor/lib/browser';
 import { FileSearchService } from '@theia/file-search/lib/common/file-search-service';
-import { FileService } from '@theia/filesystem/lib/browser/file-service';
 import { WorkspaceService } from '@theia/workspace/lib/browser/workspace-service';
 import URI from '@theia/core/lib/common/uri';
-import { COOKLANG_LANGUAGE_ID, ReportTemplates, BuiltInReportTemplate } from '../common';
-import { ReportWidget, ReportWidgetOptions, REPORT_WIDGET_ID, createReportWidgetId } from './report-widget';
+import { ReportTemplates, BuiltInReportTemplate } from '../common';
+import { ReportWidgetOptions } from './report-widget-types';
+import { ReportConfigService } from './report-config-service';
+import { ReportPresenter } from './report-presenter';
 
 // ---------------------------------------------------------------------------
 // Commands
@@ -54,20 +53,8 @@ interface ReportTemplatePick {
 @injectable()
 export class ReportContribution implements CommandContribution, MenuContribution {
 
-    @inject(EditorManager)
-    protected readonly editorManager: EditorManager;
-
-    @inject(ApplicationShell)
-    protected readonly shell: ApplicationShell;
-
-    @inject(WidgetManager)
-    protected readonly widgetManager: WidgetManager;
-
     @inject(QuickPickService)
     protected readonly quickPickService: QuickPickService;
-
-    @inject(FileService)
-    protected readonly fileService: FileService;
 
     @inject(FileSearchService)
     protected readonly fileSearchService: FileSearchService;
@@ -75,12 +62,18 @@ export class ReportContribution implements CommandContribution, MenuContribution
     @inject(WorkspaceService)
     protected readonly workspaceService: WorkspaceService;
 
+    @inject(ReportConfigService)
+    protected readonly reportConfigService: ReportConfigService;
+
+    @inject(ReportPresenter)
+    protected readonly reportPresenter: ReportPresenter;
+
     // --- CommandContribution ---
 
     registerCommands(commands: CommandRegistry): void {
         commands.registerCommand(CooklangReportCommands.RENDER_REPORT, {
             execute: () => this.renderReport(),
-            isEnabled: () => this.getActiveCooklangUri() !== undefined,
+            isEnabled: () => this.reportConfigService.getActiveCooklangUri() !== undefined,
         });
     }
 
@@ -96,7 +89,7 @@ export class ReportContribution implements CommandContribution, MenuContribution
     // --- Command execution ---
 
     protected async renderReport(): Promise<void> {
-        const uri = this.getActiveCooklangUri();
+        const uri = this.reportConfigService.getActiveCooklangUri();
         if (!uri) {
             return;
         }
@@ -109,56 +102,9 @@ export class ReportContribution implements CommandContribution, MenuContribution
             templateId: template.id,
             templateLabel: template.label,
             templateUri: template.uri,
-            configJson: await this.buildConfigJson(),
+            configJson: await this.reportConfigService.buildConfigJson(),
         };
-        const widget = await this.getOrCreateReport(options);
-        if (!widget.isAttached) {
-            await this.shell.addWidget(widget, { area: 'main' });
-        }
-        this.shell.activateWidget(widget.id);
-    }
-
-    /**
-     * Resolves the recipe URI from the focused widget, the current main-area
-     * tab, or the active Cooklang text editor — in that order. `.cook` files
-     * open in preview mode by default, and the preview widget never takes DOM
-     * focus, so it is reported by `getCurrentWidget('main')` (tab selection)
-     * but never by `shell.currentWidget` (focus tracker).
-     */
-    protected getActiveCooklangUri(): URI | undefined {
-        return this.getCooklangResourceUri(this.shell.currentWidget)
-            ?? this.getCooklangResourceUri(this.shell.getCurrentWidget('main'))
-            ?? this.getActiveCooklangEditorUri();
-    }
-
-    /**
-     * Returns the widget's resource URI when it is a navigatable showing a
-     * `.cook` or `.menu` resource (text editor, recipe preview, report tab).
-     */
-    protected getCooklangResourceUri(widget: Widget | undefined): URI | undefined {
-        if (NavigatableWidget.is(widget)) {
-            const uri = widget.getResourceUri();
-            if (uri && (uri.path.ext === '.cook' || uri.path.ext === '.menu')) {
-                return uri;
-            }
-        }
-        return undefined;
-    }
-
-    /**
-     * Returns the URI of the active editor when its language is Cooklang,
-     * or `undefined` otherwise.
-     */
-    protected getActiveCooklangEditorUri(): URI | undefined {
-        const editorWidget = this.editorManager.currentEditor;
-        if (!editorWidget) {
-            return undefined;
-        }
-        const { languageId, uri } = editorWidget.editor.document;
-        if (languageId !== COOKLANG_LANGUAGE_ID) {
-            return undefined;
-        }
-        return new URI(uri);
+        await this.reportPresenter.show(options);
     }
 
     /**
@@ -237,59 +183,5 @@ export class ReportContribution implements CommandContribution, MenuContribution
                 };
             })
             .sort((a, b) => a.label.localeCompare(b.label));
-    }
-
-    /**
-     * Builds the render config from workspace conventions. Paths are sent as
-     * URI strings; the backend converts them to filesystem paths.
-     */
-    protected async buildConfigJson(): Promise<string> {
-        const config: {
-            scale: number;
-            basePath?: string;
-            aislePath?: string;
-            pantryPath?: string;
-            datastorePath?: string;
-        } = { scale: 1 };
-        const root = this.workspaceService.tryGetRoots()[0];
-        if (root) {
-            config.basePath = root.resource.toString();
-            const aisle = root.resource.resolve('config/aisle.conf');
-            const pantry = root.resource.resolve('config/pantry.conf');
-            const datastores = [root.resource.resolve('db'), root.resource.resolve('config/db')];
-            const [hasAisle, hasPantry, ...hasDatastores] = await Promise.all([
-                this.fileService.exists(aisle),
-                this.fileService.exists(pantry),
-                ...datastores.map(candidate => this.fileService.exists(candidate)),
-            ]);
-            if (hasAisle) {
-                config.aislePath = aisle.toString();
-            }
-            if (hasPantry) {
-                config.pantryPath = pantry.toString();
-            }
-            const datastore = datastores.find((candidate, index) => hasDatastores[index]);
-            if (datastore) {
-                config.datastorePath = datastore.toString();
-            }
-        }
-        return JSON.stringify(config);
-    }
-
-    /**
-     * Returns an existing report widget for (uri, template) — looked up by its
-     * widget id among the widgets created by the report factory — otherwise
-     * creates one via the widget factory. A fresh `setOptions` re-render is
-     * triggered on reuse so the report reflects the latest config.
-     */
-    protected async getOrCreateReport(options: ReportWidgetOptions): Promise<ReportWidget> {
-        const widgetId = createReportWidgetId(new URI(options.uri), options.templateId);
-        const existing = this.widgetManager.getWidgets(REPORT_WIDGET_ID)
-            .find((widget): widget is ReportWidget => widget.id === widgetId);
-        if (existing) {
-            existing.setOptions(options);
-            return existing;
-        }
-        return this.widgetManager.getOrCreateWidget<ReportWidget>(REPORT_WIDGET_ID, options);
     }
 }
